[system call] add getpid, uart_read, uart_write
