+++ Lab5/kernel/src/exception.c
@@ -1,5 +1,5 @@
 #include "exception.h"
-
+#include "thread.h" 
 #include "io.h"
 #include "mini_uart.h"
 #include "timer.h"
@@ -18,15 +18,92 @@ void sync_handler() {
   asm volatile("mrs %0, spsr_el1" : "=r"(spsr_el1));
   asm volatile("mrs %0, elr_el1" : "=r"(elr_el1));
   asm volatile("mrs %0, esr_el1" : "=r"(esr_el1));
-  //print_s("SPSR_EL1: ");
-  //print_h(spsr_el1);
-  //print_s("\n");
-  //print_s("ELR_EL1: ");
-  //print_h(elr_el1); // return address of EL1 exception
-  //print_s("\n");
-  //print_s("ESR_EL1: "); 
-  //print_h(esr_el1); // cause of the exception
-  //print_s("\n\n");
+  print_s("SPSR_EL1: ");
+  print_h(spsr_el1);
+  print_s("\n");
+  print_s("ELR_EL1: ");
+  print_h(elr_el1); // return address of EL1 exception
+  print_s("\n");
+  print_s("ESR_EL1: "); 
+  print_h(esr_el1); // cause of the exception
+  print_s("\n\n");
+}
+
+void lower_64_EL_sync_handler(uint64_t sp){
+  // exception registor
+  uint64_t spsr_el1, elr_el1, esr_el1;
+  asm volatile("mrs %0, spsr_el1" : "=r"(spsr_el1));
+  asm volatile("mrs %0, elr_el1" : "=r"(elr_el1));
+  asm volatile("mrs %0, esr_el1" : "=r"(esr_el1));
+  asm volatile("mrs %0, esr_el1" : "=r"(esr_el1));
+  // https://developer.arm.com/documentation/ddi0595/2021-06/AArch64-Registers/ESR-EL1--Exception-Syndrome-Register--EL1-
+  int ec = (esr_el1 >> 26) & 0x3f; // exception class
+  int iss = esr_el1 & ((1 << 25) - 1); // [24:0], instruction specific syndrome {svc <iss>}
+
+  if(ec == 0b010101){
+    // according the lab spec, the number of the systme call is store in x8
+    trap_frame_t *trap_frame = (trap_frame_t *)sp;
+    int sys_call_num;
+    asm volatile("mov %0, x8" : "=r"(sys_call_num));
+
+    print_s("system call number: ");
+    print_i(sys_call_num);
+    print_s("\r\n");
+
+    // print content of the trap frame
+    //for(int i = 0; i< 10; i++){
+    //  print_s("trap frame ");
+    //  print_i(i);
+    //  print_s(", ");
+    //  print_i(trap_frame->x[i]);
+    //  print_s("\r\n");
+    //}
+
+    delay(100000000);
+
+    if(sys_call_num == 0){ // int getpid()
+      uint32_t pid = get_current()->pid;
+      asm volatile("mov x0, %0" : "=r"(pid)); // function reture value
+    }else if(sys_call_num == 1){ // uartread(char buf[], size_t size)
+      char *str = (char *)(trap_frame->x[0]);
+      uint32_t size = (uint32_t)(trap_frame->x[1]);
+      size = uart_gets(str, size);
+      trap_frame->x[0] = size;
+    }else if(sys_call_num == 2){ // uartread(char buf[], size_t size)
+      char* str = trap_frame->x[0];
+      uart_puts(str);
+    }else{
+      print_s("unhandled system call number\r\n");
+    }
+  }
+  //if(ec == 0b010101){
+  //  trap_frame_t *trap_frame = (trap_frame_t *)sp;
+  //  int sys_call_num = trap_frame->x[8];
+  //  if(sys_call_num == 0){
+  //    print_s("x8: ");
+  //    print_i(trap_frame->x[8]);
+  //    print_s("\r\n");
+  //    print_s("get pid\r\n");
+  //    uint32_t pid = get_current()->pid;
+  //    trap_frame->x[0] = pid;
+  //  }else  if(sys_call_num <10){
+  //    print_s("x8: ");
+  //  print_i(trap_frame->x[8]);
+  //  print_s("\r\n");
+  //  }
+  //
+  //}
+  //
+//
+  //
+  ////for(int i = 0; i<32; i++){
+  ////  print_s("x");
+  ////  print_i(i);
+  ////  print_s(": ");
+  ////  print_i()
+  ////  print_s("x")
+  ////  print_s("x")
+  ////}
 }
 
 void el0_to_el1_irq_handler() {
