+++ Lab5/kernel/src/mini_uart.c
@@ -64,6 +64,19 @@ char uart_getc() {
   return r == '\r' ? '\n' : r;
 }
 
+uint32_t uart_gets(char *buf, uint32_t size) {
+  for (int i = 0; i < size; ++i) {
+    buf[i] = uart_getc();
+    uart_send(buf[i]);
+    if (buf[i] == '\n' || buf[i] == '\r') {
+      uart_send('\r');
+      buf[i] = '\0';
+      return i;
+    }
+  }
+  return size;
+}
+
 void uart_puts(char *s) {
   while (*s) {
     // convert '\n' to carrige '\r' and '\n'
